oll: Rewrite \loadModule (WIP) -> Scheme modules
THis is a preliminary step towards supporting Scheme modules,
trying to harmonize the process of determining the to-be-included
file.

However, it doesn't really work yet: The module file is included
(and if one adds a (ly:message) in it this is produced, but
it seems the defined commands are not available afterwards.